Hooking overrides.

+++ Utilities/Encoding/Widestring.hpp
@@ -12,16 +12,16 @@
 #include <Windows.h>
 [[nodiscard]] inline std::string toNarrow(const std::wstring_view Input)
 {
-    const auto Size = WideCharToMultiByte(CP_UTF8, 0, Input.data(), Input.size(), NULL, 0, NULL, FALSE);
+    const auto Size = WideCharToMultiByte(CP_UTF8, 0, Input.data(), (int)Input.size(), NULL, 0, NULL, FALSE);
     const auto Buffer = (char *)alloca(Size * sizeof(char) + sizeof(char));
-    WideCharToMultiByte(CP_UTF8, 0, Input.data(), Input.size(), Buffer, Size, NULL, FALSE);
+    WideCharToMultiByte(CP_UTF8, 0, Input.data(), (int)Input.size(), Buffer, Size, NULL, FALSE);
     return std::string(Buffer, Size);
 }
 [[nodiscard]] inline std::wstring toWide(const std::string_view Input)
 {
-    const auto Size = MultiByteToWideChar(CP_UTF8, 0, Input.data(), Input.size(), NULL, 0);
+    const auto Size = MultiByteToWideChar(CP_UTF8, 0, Input.data(), (int)Input.size(), NULL, 0);
     const auto Buffer = (wchar_t *)alloca(Size * sizeof(wchar_t) + sizeof(wchar_t));
-    MultiByteToWideChar(CP_UTF8, 0, Input.data(), Input.size(), Buffer, Size);
+    MultiByteToWideChar(CP_UTF8, 0, Input.data(), (int)Input.size(), Buffer, Size);
     return std::wstring(Buffer, Size);
 }
